Bug 1407710, 1407707, 1407709, 1407702 — Add ScreenActions, ScreenState and UserState to ScreenGraph.
https://github.com/mozilla-mobile/firefox-ios/issues/3272

[ba0457f] Bug 1407710 — Add better tidy up of the backstack.
[e420368] Bug 1407710 — Uncontroversial cleanup
[baa1fcc] Bug 1407710 — Migrate to new onEnter methods.
[bde8f5b] Bug 1407710 — Add navigator.plan()
[61709d2] Bug 1407710 — Split onEnter into onEnter and onEnterWaitFor.

This lets us add a conditional waitFor.

screenState.onEnterWaitFor(element:, if: "")
[04cc405] Bug 1407710 — Reduce visibility of test identifiers

This was causing clashes with FxScreenGraph.
[5631de4] Bug 1407710 — Add back() method, press() method, synchronizeWithUserState() method
[a2f9332] Bug 1407710 — Add extra methods to expose new actions and conditional edges to all gestures.

Improve error reporting.
[c03a87e] Bug 1407707 — Add conditional edges to the screen graph.

This lets us define edges (including actions) as conditional on the current userstate.

If the userState changes (even mid `goto`), the conditional edges are re-evaluated, and the navigator is re-routed.

Unfortunately, this commit adds mutable state to the graph, which means that there is now a one-to-one navigator-to-screengraph mapping (before you could have multiple navigators from the same instance of the ScreenGraph. This is almost certainly not worth worrying about.
[e73fc54] Bug 1407706 — More testing of actions
[9bd748e] Bug 1407706 — More testing of actions
[904c492] Bug 1407706 — Add toggleOn and toggleOff actions.

These depend on the userState being up to date.
[868d2e5] Bug 1407706 — Move enter and leave into its own private extension.
[13bdbd4] Bug 1407706 — Ensure actions always end in screen states.
[327f9b5] Bug 1407706 — Actions can be added from within screen states.
[a012a0e] Bug 1407706 — Refactor navigator.goto() into a enter and leave methods.

This also changes currentScene to be using the new generalization of GraphNode.
[059a011] Bug 1407706 — Start ScreenActionNodes

Refactor ScreenGraphNode to ScreenStateNode and GraphNode.
[6f994c0] Bug 1407709 — Add predicate tests for entry criteria into nodes.
[dc3063e] Bug 1407702 — Add onEnter user state recorders.

Unify how we look at the state of the app.
[3558577] Bug 1407702 — Add failing tests for onEnter state changes
[84cbccf] Bug 1407702 — Add UserState to ScreenGraph.

This is the basis of adding actions and predicates to the screengraph. As neither of these exist, this is not very useful.